requirements.txt
boto3==1.5.6
orator==0.9.7
PyJWT==1.5.3
PyMySQL==0.8.0
pytest==3.3.1
pytest-watch==4.1.0
python-slugify==1.2.4
PyYAML==3.12
